Backport #56117 to release/9.8 - plugins/woocommerce-blocks/assets/js/blocks/product-gallery/types.ts
