+++ plugins/woocommerce-blocks/assets/js/blocks/product-gallery/frontend.ts
@@ -7,19 +7,10 @@ import {
 	getElement,
 } from '@wordpress/interactivity';
 
-export interface ProductGalleryContext {
-	// It's an actual image number, not an index, hence one-based!
-	selectedImageNumber: number;
-	imageId: string;
-	imageIds: string[];
-	isDialogOpen: boolean;
-	productId: string;
-	disableLeft: boolean;
-	disableRight: boolean;
-	touchStartX: number;
-	touchCurrentX: number;
-	isDragging: boolean;
-}
+/**
+ * Internal dependencies
+ */
+import type { ProductGalleryContext } from './types';
 
 const getContext = ( ns?: string ) =>
 	getContextFn< ProductGalleryContext >( ns );
