+++ plugins/woocommerce-blocks/assets/js/blocks/product-gallery/types.ts
@@ -10,3 +10,17 @@ export interface ProductGallerySettingsProps {
 		attributes: Partial< ProductGalleryBlockAttributes >
 	) => void;
 }
+
+export interface ProductGalleryContext {
+	// It's an actual image number, not an index, hence one-based!
+	selectedImageNumber: number;
+	imageId: string;
+	imageIds: string[];
+	isDialogOpen: boolean;
+	productId: string;
+	disableLeft: boolean;
+	disableRight: boolean;
+	touchStartX: number;
+	touchCurrentX: number;
+	isDragging: boolean;
+}
